fix: quicksave bobinage

+++ app/src/main/java/com/example/applicationstb/localdatabase/RemontageCCEntity.kt
@@ -25,8 +25,8 @@ data class RemontageCCEntity (
     var collageRoulementPorteeFlasque: Int?,
     var verificationFixationCouronne: Boolean?,
     var verificationIsolementPorteBalais: Boolean?,
-    var isolementPorteBalaisV: Int?,
-    var isolementPorteBalaisOhm: Int?,
+    var isolementPorteBalaisV: Float?,
+    var isolementPorteBalaisOhm: Float?,
     // essais dynamiques
     var tensionStator:Boolean?,
     var tensionStatorU:Float?,
